Generalise stderr label details to tabular form
We produce a little overview of the labels processed as a tree which
prints to stderr, but this doesn't generally get looked at past the
processing stage. Switch to a general mechanism and provide an
alternative that creates tabular output that might be useful for
validation later on.

Also drive-by tidy up the omit_layers logic a little.