Correctly handle dataCollections for new labels
Inside the labels script, the logical fork for handling the new format
exported does not correctly cope with the real world shape of the
dataCollections property. This doesn't match the currently documented
schema either, but has not changed from the old format, so can be
updated to just have the same handling as the existing code for the
previous format. Also rename some variables to make the pieces here
clearer.